came up with a satisfying solution for the throwing
it started with this solution 
https://gamedev.stackexchange.com/questions/61301/how-to-implement-throw-curve-with-virtual-height-in-a-2d-side-view-game?noredirect=1&lq=1
it was great for long distances, however short-ranged it barely went up. instead i tweaked it so that it would be time dependent rather than distance dependent. every shot would take 1 or 2 seconds regardless of how short or how far, this made the parabolic arc prominent regardless of where the target was and that's what i wanted :).